requirements.txt
annotated-types==0.5.0
anyio==3.7.1
attrs==23.2.0
bcrypt==4.1.3
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
exceptiongroup==1.2.1
fastapi==0.68.0
h11==0.12.0
httpcore==0.13.7
httpx==0.19.0
idna==3.7
importlib-metadata==6.7.0
iniconfig==2.0.0
mysql-connector-python==8.0.33
packaging==24.0
passlib==1.7.4
pluggy==1.2.0
protobuf==3.20.3
py==1.11.0
pydantic==1.8.2
pydantic_core==2.14.6
pytest==7.1.3
python-multipart==0.0.8
requests==2.31.0
rfc3986==1.5.0
sniffio==1.3.1
starlette==0.14.2
tomli==2.0.1
typing_extensions==4.7.1
urllib3==1.26.6
uvicorn==0.22.0
zipp==3.15.0